requirements.txt
pymongo
langchain
pandas
pypdf
gdown
langchain_google_genai
langchain-community
python-dotenv